feat: ReservedDate 엔티티 생성
