--- backend/src/main/java/com/team3/airdnd/accommodation/domain/Reservation.java
@@ -1,45 +1,0 @@
-package com.team3.airdnd.accommodation.domain;
-
-import com.team3.airdnd.user.domain.User;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Reservation {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "guest_id", nullable = false)
-    private User guest;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "accommodation_id", nullable = false)
-    private Accommodation accommodation;
-
-    @Column(name = "check_in", nullable = false)
-    private LocalDate checkIn;
-
-    @Column(name = "check_out", nullable = false)
-    private LocalDate checkOut;
-
-    @Column(name = "guest_count", nullable = false)
-    private Integer guestCount;
-
-    @Enumerated(EnumType.STRING)
-    private Status status;
-
-    @Column(name = "total_price", nullable = false)
-    private Long totalPrice;
-
-    public enum Status {
-        PENDING, CONFIRMED, CANCELLED
-    }
-}
