+++ backend/src/main/java/com/team3/airdnd/accommodation/repository/ReservationRepository.java
@@ -2,7 +2,7 @@ package com.team3.airdnd.accommodation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team3.airdnd.accommodation.domain.Reservation;
+import com.team3.airdnd.reservation.domain.Reservation;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 	boolean existsByAccommodationId(Long accommodationId);
