+++ backend/src/main/java/com/team3/airdnd/accommodation/repository/ReviewRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.team3.airdnd.accommodation.domain.Review;
 import com.team3.airdnd.accommodation.dto.ReviewDto;
+import com.team3.airdnd.review.domain.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
