+++ backend/src/main/java/com/team3/airdnd/accommodation/service/AccommodationService.java
@@ -19,8 +19,6 @@ import com.team3.airdnd.accommodation.domain.Amenity;
 import com.team3.airdnd.accommodation.domain.AmenityType;
 import com.team3.airdnd.accommodation.domain.QAccommodation;
 import com.team3.airdnd.accommodation.domain.QAddress;
-import com.team3.airdnd.accommodation.domain.QReservation;
-import com.team3.airdnd.accommodation.domain.Reservation;
 import com.team3.airdnd.accommodation.dto.AccommodationRequestDto;
 import com.team3.airdnd.accommodation.dto.AccommodationResponseDto;
 import com.team3.airdnd.accommodation.dto.HostAccommodationQueryDto;
@@ -35,6 +33,8 @@ import com.team3.airdnd.accommodation.repository.ReservationRepository;
 import com.team3.airdnd.accommodation.repository.ReviewRepository;
 import com.team3.airdnd.global.exception.CommonException;
 import com.team3.airdnd.global.exception.ErrorCode;
+import com.team3.airdnd.reservation.domain.QReservation;
+import com.team3.airdnd.reservation.domain.Reservation;
 import com.team3.airdnd.storedFile.StoredFileService;
 import com.team3.airdnd.storedFile.domain.StoredFile;
 import com.team3.airdnd.storedFile.repository.StoredFileRepository;
