+++ backend/src/main/java/com/team3/airdnd/chat/domain/ChatRoom.java
@@ -1,11 +1,24 @@
 package com.team3.airdnd.chat.domain;
 
-import com.team3.airdnd.accommodation.domain.Reservation;
-import com.team3.airdnd.user.domain.User;
-import jakarta.persistence.*;
-import lombok.*;
-
 import java.time.LocalDateTime;
+
+import com.team3.airdnd.reservation.domain.Reservation;
+import com.team3.airdnd.user.domain.User;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.PrePersist;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Builder
@@ -13,33 +26,33 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChatRoom {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "host_id", nullable = false)
-    private User host;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "host_id", nullable = false)
+	private User host;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "guest_id", nullable = false)
-    private User guest;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "guest_id", nullable = false)
+	private User guest;
 
-    @OneToOne
-    @JoinColumn(name = "reservation_id")
-    private Reservation reservation;
+	@OneToOne
+	@JoinColumn(name = "reservation_id")
+	private Reservation reservation;
 
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
+	@Column(name = "created_at", nullable = false)
+	private LocalDateTime createdAt;
 
-    @Column(name = "last_message", length=500)
-    private String lastMessage;
+	@Column(name = "last_message", length = 500)
+	private String lastMessage;
 
-    @Column(name = "last_message_at")
-    private LocalDateTime lastMessageAt;
+	@Column(name = "last_message_at")
+	private LocalDateTime lastMessageAt;
 
-    @PrePersist
-    public void prePersist() {
-        this.createdAt = LocalDateTime.now();
-    }
+	@PrePersist
+	public void prePersist() {
+		this.createdAt = LocalDateTime.now();
+	}
 }
